+++ lib/scss_lint/options.rb
@@ -102,6 +102,10 @@ module SCSSLint
         @options[:show_linters] = true
       end
 
+      parser.on('--[no-]color', 'Force output to be colorized') do |color|
+        @options[:color] = color
+      end
+
       parser.on_tail('-h', '--help', 'Display help documentation') do
         @options[:help] = parser.help
       end