+++ spec/scss_lint/options_spec.rb
@@ -30,5 +30,23 @@ describe SCSSLint::Options do
         expect { subject }.to raise_error SCSSLint::Exceptions::InvalidCLIOption
       end
     end
+
+    context 'color' do
+      describe 'manually on' do
+        let(:args) { ['--color'] }
+
+        it 'sets the `color` option to true' do
+          subject.should include color: true
+        end
+      end
+
+      describe 'manually off' do
+        let(:args) { ['--no-color'] }
+
+        it 'sets the `color option to false' do
+          subject.should include color: false
+        end
+      end
+    end
   end
 end